rename reconnect_handler to subscription_handler
